+++ web/module.go
@@ -1,0 +1,8 @@
+package web
+
+import "go.uber.org/fx"
+
+var Module = fx.Provide(
+	newFiberMiddlewareRegistry,
+	newFiber,
+)